Support per-device Low Power Mode configuration
Broadcom's Bluedroid's libbt-vendor implementation provides a set of
configurable values for the sleep_mode initialization. For example, some
devices have an active IRQ polarity set to 0 instead of the default 1.
In order to properly configure the BT chip, this patch exposes a set
of known LPM configurations choosable with LPM_DEVICE macro.